+++ Models/TickerData.cs
@@ -14,9 +14,9 @@ namespace OoplesFinance.StockIndicators.Models;
 public class TickerData : ITickerData
 {
     public DateTime Date { get; set; }
-    public decimal Open { get; set; }
-    public decimal High { get; set; }
-    public decimal Low { get; set; }
-    public decimal Close { get; set; }
-    public decimal Volume { get; set; }
+    public double Open { get; set; }
+    public double High { get; set; }
+    public double Low { get; set; }
+    public double Close { get; set; }
+    public double Volume { get; set; }
 }
